Allow Dynamic Lists for Packets
There are packets coming up which have any number of repeating
substructure fields in their payload.  Prior to this the only
list we had was for characters at the select screen; which was
fixed at 10.  It should be noted that currently this dynamic
selection *must* be the last field in a packet.  If any more
packets show up that have additional fields past a list we'll
need to go back and revamp how the packet sytem works to allow
for partial deconstruction of a packet in the intial match of
a binary.